Extract OSCALRoles to seperate file
